requirements.txt
# Google Gemini API (nova dependência para Gemini 2.0 Flash)
google-genai>=0.3.0

# Dependências existentes
google-generativeai
python-dotenv
requests
Pillow
gtts
moviepy>=2.2.1
ffmpeg-python
opencv-python

# ElevenLabs TTS (nova dependência para narração avançada)
elevenlabs>=0.2.26

# Flask para API backend do frontend web
Flask>=2.3.0
Flask-CORS>=4.0.0

# Removidas temporariamente: coqui-tts (problemas de compatibilidade)

# Removidas: google-cloud-aiplatform e vertexai (substituídas por google-genai)
